pyproject.toml
[tool.poetry]
name = "python-test"
version = "0.1.0"
description = ""
authors = ["Umbreella <draginvadim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
uvicorn = { extras = ["standard"], version = "^0.23.2" }
sqlalchemy = "^2.0.20"
psycopg-binary = "^3.1.10"
asyncpg = "^0.28.0"
alembic = "^1.11.3"
pydantic-settings = "^2.0.3"
httpx = "^0.25.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
flake8 = "^6.1.0"
flake8-quotes = "^3.3.2"
flake8-broken-line = "^1.0.0"
flake8-commas = "^2.1.0"
isort = "^5.12.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pytest-postgresql = "^5.0.0"
pytest-cov = "^4.1.0"
faker = "^19.10.0"


[tool.isort]
skip = ["*/__init__.py"]
multi_line_output = 3
include_trailing_comma = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
